Header Section
Finished the header section after some trouble with
flexbox and the navbar at the top. I wanted the
spacing to be the same as the image below the nav bar
but for some reason I couldn't figure out how to do it.
I think justify-content was how I fixed it, adding both
justify-content: space-between; and having similar codes
to the divs.