Information Center
Added the information center which mimics TOP's "some random information"
section. Pretty easy for me to figure out for the most part. I removed the p
elements from being under the img section, and instead, created a div stricly for
the images and for the p elements. Then I just added classes to them and
added the same flex settings on them. One thing that does bother me is the quote
not lining up with the images width wise, but I wont stress over that for now.