Bottom Container
Took me longer than I expected. I had a lot of trouble with positioning
of the content in the blue part. Also, the width to 50% has
a different color for the sides there? I need to fix that.
I also fixed the hover button at the top to have an effect, similar
to the hover effect on the bottom container's button.